Extract shared lock-free block helpers to cut SonarCloud duplication.
The MPSC and MPMC block queues carried byte-identical graveyard and
block-linking infrastructure (allocate_block, retire_block,
link_next_block), which tripped SonarCloud's duplicated-lines gate.

Move these three routines into a new templated header,
queue/detail/block_ops.hpp, as non-virtual free function templates that
inline to the same machine code, preserving the exact TSAN-validated
memory ordering. The intentional MPSC/MPMC differences (Block layout,
liveness model, consumer logic) remain in their respective headers. Drop
the now-unused <memory> include from both queue headers.
